Change package of Rx interactors package
