+++ example/src/main/java/com/thefuntasty/mvvmsample/App.kt
@@ -1,7 +1,7 @@
 package com.thefuntasty.mvvmsample
 
 import android.util.Log
-import com.thefuntasty.interactors.InteractorErrorHandler
+import com.thefuntasty.mvvm.usecases.UseCaseErrorHandler
 import com.thefuntasty.mvvmsample.injection.DaggerApplicationComponent
 import dagger.android.AndroidInjector
 import dagger.android.support.DaggerApplication
@@ -11,7 +11,7 @@ class App : DaggerApplication() {
     override fun onCreate() {
         super.onCreate()
 
-        InteractorErrorHandler.globalOnErrorLogger = { error ->
+        UseCaseErrorHandler.globalOnErrorLogger = { error ->
             Log.d("Interactor error", "$error")
         }
     }
