+++ example/src/main/java/com/thefuntasty/mvvmsample/domain/StateUseCase.kt
@@ -1,11 +1,11 @@
 package com.thefuntasty.mvvmsample.domain
 
-import com.thefuntasty.interactors.interactors.BaseMayber
+import com.thefuntasty.mvvm.usecases.base.BaseMayber
 import io.reactivex.Maybe
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
-class StateInteractor @Inject constructor() : BaseMayber<Boolean, Boolean>() {
+class StateUseCase @Inject constructor() : BaseMayber<Boolean, Boolean>() {
 
     companion object {
         private const val DELAY_MS = 3000L
