+++ example/src/main/java/com/thefuntasty/mvvmsample/ui/login/activity/LoginViewModel.kt
@@ -1,6 +1,6 @@
 package com.thefuntasty.mvvmsample.ui.login.activity
 
-import com.thefuntasty.interactors.BaseRxViewModel
+import com.thefuntasty.mvvm.usecases.BaseRxViewModel
 import javax.inject.Inject
 
 class LoginViewModel @Inject constructor() : BaseRxViewModel<LoginViewState>() {
