+++ interactors/src/main/java/com/thefuntasty/mvvm/usecases/BaseRxViewModel.kt
@@ -1,8 +1,8 @@
-package com.thefuntasty.interactors
+package com.thefuntasty.mvvm.usecases
 
 import androidx.lifecycle.LiveData
-import com.thefuntasty.interactors.disposables.DisposablesOwner
-import com.thefuntasty.interactors.interactors.BaseObservabler
+import com.thefuntasty.mvvm.usecases.disposables.DisposablesOwner
+import com.thefuntasty.mvvm.usecases.base.BaseObservabler
 import com.thefuntasty.mvvm.BaseViewModel
 import com.thefuntasty.mvvm.ViewState
 import io.reactivex.disposables.CompositeDisposable
