+++ interactors/src/main/java/com/thefuntasty/mvvm/usecases/base/BaseCompletabler.kt
@@ -1,18 +1,18 @@
-package com.thefuntasty.interactors.interactors
+package com.thefuntasty.mvvm.usecases.base
 
 import io.reactivex.Completable
 
 /**
  * Base interactor which wraps [Completable]. Instance of this
  * interactor can be simply executed in cooperation with
- * [com.thefuntasty.interactors.disposables.CompletableDisposablesOwner] interface.
+ * [com.thefuntasty.usecases.disposables.CompletableDisposablesOwner] interface.
  *
  * Wrapped stream is subscribed on [io.reactivex.schedulers.Schedulers.io] and
  * observed on [io.reactivex.android.schedulers.AndroidSchedulers.mainThread]
  * by default. You may override these through [workScheduler] and
  * [resultScheduler] respectively.
  */
-abstract class BaseCompletabler<ARGS> : BaseInteractor() {
+abstract class BaseCompletabler<ARGS> : BaseUseCase() {
 
     /**
      * Prepares whole wrapped [Completable] Rx stream. This method does not
