+++ interactors/src/main/java/com/thefuntasty/mvvm/usecases/base/BaseObservabler.kt
@@ -1,18 +1,18 @@
-package com.thefuntasty.interactors.interactors
+package com.thefuntasty.mvvm.usecases.base
 
 import io.reactivex.Observable
 
 /**
  * Base interactor which wraps [Observable]. Instance of this
  * interactor can be simply executed in cooperation with
- * [com.thefuntasty.interactors.disposables.ObservableDisposablesOwner] interface.
+ * [com.thefuntasty.usecases.disposables.ObservableDisposablesOwner] interface.
  *
  * Wrapped stream is subscribed on [io.reactivex.schedulers.Schedulers.io] and
  * observed on [io.reactivex.android.schedulers.AndroidSchedulers.mainThread]
  * by default. You may override these through [workScheduler] and
  * [resultScheduler] respectively.
  */
-abstract class BaseObservabler<ARGS, T> : BaseInteractor() {
+abstract class BaseObservabler<ARGS, T> : BaseUseCase() {
 
     /**
      * Prepares whole wrapped [Observable] Rx stream. This method does not
