+++ interactors/src/test/java/com/thefuntasty/mvvm/usecases/base/RxSchedulerRule.java
@@ -1,4 +1,4 @@
-package com.thefuntasty.interactors.base;
+package com.thefuntasty.mvvm.usecases.base;
 
 import io.reactivex.Scheduler;
 import io.reactivex.android.plugins.RxAndroidPlugins;
